+++ src/Lib/git.ts
@@ -10,7 +10,18 @@ export async function machineHasGit() {
 	}
 }
 
+export async function throwErrorIfGitNotInstalled() {
+	if (!(await machineHasGit())) {
+		throw new Error('Git is not installed on this machine. Please install Git to use this feature.');
+	}
+}
+
 export async function initGitRepository(relativePath: string) {
 	const absolutePath = path.resolve(relativePath);
 	await executeChildProcess(`git init "${absolutePath}"`, false);
+}
+
+export async function cloneGitRepository(gitUrl: string, targetDir: string) {
+	const absolutePath = path.resolve(targetDir);
+	await executeChildProcess(`git clone "${gitUrl}" "${absolutePath}"`, false);
 }