+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- New command `sos custom-script generate` to generate local repository with Custom Script boilerplate
+
 ## [2.1.0] - 2025-04-08
 ### Added
 - Added option to support Rspack for generated applets